Begin implementing ptrdiff_t typing for correctness
There are currently many areas that implicitly treat ssize_t
and ptrdiff_t as the same type, which isn't necessarily correct
(though really only from a pedantic standpoint). Now begins the
phase consisting of substituting ssize_t for ptrdiff_t where
it's appropriate. (Also, the format specifier for ptrdiff_t
is '%td', as already documented by the AT&T devs in the
SFIO man page.)

+++ src/cmd/ksh93/bltins/getopts.c
@@ -42,13 +42,13 @@ static int infof(Opt_t* op, Sfio_t* sp, const char* s, Optdisc_t* dp)
 	if(nv_search(s,sh.fun_tree,0))
 #endif /* SHOPT_NAMESPACE */
 	{
-		ssize_t savtop = stktell(stkp);
+		ptrdiff_t savtop = stktell(stkp);
 		void *savptr = stkfreeze(stkp,0);
 		sfputc(stkp,'$');
 		sfputc(stkp,'(');
 		sfputr(stkp,s,')');
 		sfputr(sp,sh_mactry(stkfreeze(stkp,1)),-1);
-		stkset(stkp,savptr,(size_t)savtop);
+		stkset(stkp,savptr,savtop);
 	}
 	return 1;
 }
