+++ src/cmd/ksh93/sh/fault.c
@@ -469,7 +469,7 @@ int sh_trap(const char *trap, int mode)
 	int	was_verbose = sh_isstate(SH_VERBOSE);
 	char	was_no_trapdontexec = !sh.st.trapdontexec;
 	char	save_chldexitsig = sh.chldexitsig;
-	ssize_t staktop = stktell(sh.stk);
+	ptrdiff_t staktop = stktell(sh.stk);
 	void	*savptr = stkfreeze(sh.stk,0);
 	struct	checkpt buff;
 	Fcin_t	savefc;
@@ -514,7 +514,7 @@ int sh_trap(const char *trap, int mode)
 		sh.intrap_exit_n = 0;
 	else
 		sh.exitval = savxit;
-	stkset(sh.stk,savptr,(size_t)staktop);
+	stkset(sh.stk,savptr,staktop);
 	fcrestore(&savefc);
 	if(was_history)
 		sh_onstate(SH_HISTORY);
